fix(installer): address Gemini PR #225 review (CWE-377 temp names, DRY displayName)
- Replace System.currentTimeMillis() temp file/dir names with UUID.randomUUID()
  in AppAnalyzerImpl (analysis bundle/apk) and InstallerRepositoryImpl (install
  bundle + root/shizuku/dhizuku temp dirs): unpredictable and collision-free
  even under concurrent operations (CWE-377).
- Extract the duplicated private displayNameOf() into a shared
  Uri.getDisplayName(context) extension (util/UriExtensions.kt), used from both
  AppAnalyzerImpl and InstallerRepositoryImpl.
